tests: add confidence threshold -- tests now pass
Add a confidence threshold (currently 30% confidence): the test only
checks for correct identification when the person ID is found with a
confidence > threshold.

Currently, all faces in the SinglePersonMultipleFile are correctly
identified (eg, with a confidence > threshold), and 91.8% of the faces
are correctly identified in the Multiperson test (0 false-positive and
56 true negative -- ie faces with recognition confidence < threshold --
out of tested 677 faces belonging to 6 different people).

While here, skip frames in the multi-person test to accelerate the test.